Separated index.jsx into index.jsx and app.jsx. Changed DB seed schema. DB seed schema changes =  ISBN and relatedISBN are type string instead of number since it can sometimes start with a zero.
