+++ server/index.js
@@ -10,8 +10,6 @@ app.use(express.urlencoded({extended: false}));
 // Retrieve dummy seed data
 app.get('/getBooks/:relatedIsbn', (req, res) => {
   db.getData(req, (err, data) => {
-// app.get('/getBooks/', (req, res) => {
-//   db.getData((err, data) => {
     if (err) {
       res.status(500).send(err);
     } else {
